# api_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Statistics Service - gostat. A microservice-based service, written in Golang and TypeScript.",
        "title": "Gostat",
        "contact": {
            "name": "Nikita Khvatov",
            "url": "https://khvat.pro",
            "email": "nik19ta.me@gmail.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "https://github.com/nikkhvat/gostat/blob/master/LICENSE"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/stats/visit/{session}": {
            "patch": {
                "description": "Endpoint to extend the duration of an active visit session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Extend the visit session duration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.SuccessVisitExtendResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorVisitExtendResponse"
                        }
                    }
                }
            }
        },
...
        "/stats/get/visits": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves visits data for a specific application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Retrieve visits for an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "app",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved data. The structure of the data depends on the application.",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Example: {\\\"error\\\": true, \\\"detail\\\": \\\"Detailed error message\\\"}",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorGetVisitResponse"
                        }
                    }
                }
            }
        },
        "/stats/set/visit/{app}": {
            "post": {
                "description": "Endpoint for registering a new visit or extending an existing session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Register a new visit or extend an existing session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Visit Details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.VisitData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.SuccessSetVisitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorSetVisitResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "http.ErrorGetVisitResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detailed error message\nexample: \"Detailed error message\"",
                    "type": "string"
                },
                "error": {
                    "description": "Indicates an error occurred\nexample: true",
                    "type": "boolean"
                }
            }
        },
        "http.ErrorSetVisitResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detailed error message\nexample: \"Detailed error message\"",
                    "type": "string"
                },
                "error": {
                    "description": "Indicates an error occurred\nexample: true",
                    "type": "boolean"
                }
            }
        },
        "http.ErrorVisitExtendResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detailed error message\nexample: \"Detailed error message\"",
                    "type": "string"
                },
                "error": {
                    "description": "Indicates an error occurred\nexample: true",
                    "type": "boolean"
                }
            }
        },
...
        "http.SuccessSetVisitResponse": {
            "type": "object",
            "properties": {
                "session": {
                    "description": "Session ID for the visit\nexample: \"session_id\"",
                    "type": "string"
                },
                "successfully": {
                    "description": "Indicates successful visit setting or extending\nexample: true",
                    "type": "boolean"
                }
            }
        },
        "http.SuccessVisitExtendResponse": {
            "type": "object",
            "properties": {
                "successfully": {
                    "description": "Indicates successful visit extension\nexample: true",
                    "type": "boolean"
                }
            }
        },
        "http.UTMParams": {
            "type": "object",
            "properties": {
                "utm_campaign": {
                    "type": "string"
                },
                "utm_content": {
                    "type": "string"
                },
                "utm_medium": {
                    "type": "string"
                },
                "utm_source": {
                    "type": "string"
                },
                "utm_term": {
                    "type": "string"
                }
            }
        },
        "http.VisitData": {
            "type": "object",
            "properties": {
                "expired": {
                    "type": "boolean"
                },
                "hash": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "pathname": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "utm": {
                    "$ref": "#/definitions/http.UTMParams"
                }
            }
        },
...
    }
}